requirements.txt
ultralytics>=8.3.0
opencv-python>=4.8.0
numpy>=1.24
pandas>=2.0
Pillow>=9.5
matplotlib>=3.7
scikit-learn>=1.2
albumentations>=1.3.0
torchmetrics>=1.0.0
tqdm>=4.65.0
natsort>=8.4.0
wandb>=0.16.0